Changes per request from ptone
