+++ django/contrib/messages/tests/__init__.py
@@ -2,3 +2,4 @@ from django.contrib.messages.tests.cookie import CookieTest
 from django.contrib.messages.tests.fallback import FallbackTest
 from django.contrib.messages.tests.middleware import MiddlewareTest
 from django.contrib.messages.tests.session import SessionTest
+from django.contrib.messages.tests.mixins import SuccessMessageMixinTests
